# Adding DogStatsd Codec