--- a/tremor-cli/tests/bench/passthrough-dogstatsd/config.troy
+++ b/tremor-cli/tests/bench/passthrough-dogstatsd/config.troy
@@ -1,0 +1,15 @@
+define flow main
+flow
+  use tremor::pipelines;
+  use bench;
+  create connector bench from bench::bench
+  with
+    file = "data.dogstatsd.xz",
+    codec = "dogstatsd"
+  end;
+  create pipeline main from pipelines::passthrough;
+
+  connect /connector/bench to /pipeline/main;
+  connect /pipeline/main to /connector/bench;
+end;
+deploy flow main;