--- a/tremor-cli/tests/bench/passthrough-dogstatsd/tags.yaml
+++ b/tremor-cli/tests/bench/passthrough-dogstatsd/tags.yaml
@@ -1,0 +1,4 @@
+[
+    "passthrough",
+    "codec:dogstatsd"
+]