--- a/tremor-cli/tests/bench/passthrough-statsd/config.troy
+++ b/tremor-cli/tests/bench/passthrough-statsd/config.troy
@@ -1,0 +1,15 @@
+define flow main
+flow
+  use tremor::pipelines;
+  use bench;
+  create connector bench from bench::bench
+  with
+    file = "data.statsd.xz",
+    codec = "statsd"
+  end;
+  create pipeline main from pipelines::passthrough;
+
+  connect /connector/bench to /pipeline/main;
+  connect /pipeline/main to /connector/bench;
+end;
+deploy flow main;