--- a/tremor-cli/tests/bench/passthrough-statsd/tags.yaml
+++ b/tremor-cli/tests/bench/passthrough-statsd/tags.yaml
@@ -1,0 +1,4 @@
+[
+    "passthrough",
+    "codec:statsd"
+]