… demo/data/data.dogstatsd.xz
… demo/data/data.statsd.xz