@mfelsche 2022-09-13
The benchmarks in the `bench` folder are unfortunately completely out of date and we didn't clean them up yet. The new benchmarks, whose results are also shown at https://tremor.rs/benchmarks are in the folder `tremor-cli/tests/bench`.

Sorry for the confusion, we promise to take the time to clean those up.