@mfelsche 2022-09-14
Unfortunately clippy is not happy (just two little things) and our safety checker detected some unwraps and array access.
If you are 100% sure the array access is safe, put a comment like: `// ALLOW: <explanation why this is safe>` above it, if not use something like `array.get(2..).ok_or_(ErrorKind::InvalidDogStatsD)`