@mfelsche 2022-09-14
I am wondering why the throughput numbers between dogstatsd and statsd are so different. I guess the incoming events have different sizes?